+++ packages/edit/i18n/it.js
@@ -9,5 +9,5 @@ i18n.map('it', {
 	btn_ren: "Rinomina",
 	btn_reload: "Ricarica",
 
-	label_editplaces: "Luoghi creati"
+	label_editplaces: "Modifiche"
 });