+++ packages/lib/lib/Autosize.js
@@ -1,0 +1,2 @@
+
+Autosize = autosize;