+++ packages/core-ui/client/views/panels/place.js
@@ -1,0 +1,11 @@
+
+/*
+//TODO fix auto close sidebar before uncomment
+when click on popupt title sidebar not shown
+
+Template.panelPlace_ui_btns.events({
+	'click .place-btn-map': function(e, tmpl) {
+		e.preventDefault();
+		tmpl.data.showLoc();
+	}
+});*/