+++ packages/core-ui/settings.js
@@ -4,6 +4,8 @@ Kepler.settings.public.templates = {
 	//TODO refact.. templates without placeholder
 	"tabPlace_ui_info": { "show": true},
 	"tabPlace_ui_hist": { "show": true},
+	"tabPlace_ui_nearby": { "show": true, order:-10},
+	
 	"tabUser_ui_info": { "show": true},
 	"tabUser_ui_friends": { "show": true},
 	"tabUser_ui_hist": { "show": true},
@@ -36,11 +38,11 @@ Kepler.settings.public.templates = {
 	},
 	"panelPlace": {
 		"panelPlace_ui_btn_checkin": { order:-10 },
-		"panelPlace_ui_btns": { order:-10 },
-		"panelPlace_ui_info": { order:-10 }
+		"panelPlace_ui_btns": { order:-10 }
 	},
 	"tabPlace": {
-		"tabPlace_ui_info": { order:-10 }
+		"tabPlace_ui_info": { order:-10 },
+		"tabPlace_ui_nearby": { order:-10 },
 	},			
 	"panelUser": {
 		"panelUser_ui_bio": { order:-20 },